- broke down SettingsSidebar to other componenets - move timecalcuation withsidebar compoenents
