+++ public/react/Configure/SettingsSidebar/components/Branding/Branding.tsx
@@ -1,0 +1,29 @@
+import React from "react";
+
+interface BrandingProps {
+  showSidebarBranding: boolean;
+}
+
+const Branding: React.FC<BrandingProps> = ({ showSidebarBranding }) => (
+showSidebarBranding ? (
+  <div className="flex gap-2 pt-4">
+    <div className="flex-none pt-1">
+      <img src="./images/eggbert-light-minimum.svg"/>
+    </div>
+    <div className="flex-auto grow items-baseline leading-4">
+      <div className="color-gray-900 underline-on-hover bitovi-font-poppins font-bold">
+        <a href="https://github.com/bitovi/jira-timeline-report" target="_blank">
+          Status Reports
+        </a>
+      </div>
+      <div className="bitovi-poppins text-neutral-100 text-sm">
+        <a href="https://www.bitovi.com/services/agile-project-management-consulting" target="_blank">
+          by Bitovi
+        </a>
+      </div>
+    </div>
+  </div>
+) : null
+);
+
+export default Branding;
