+++ public/react/Configure/SettingsSidebar/components/Branding/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./Branding";
