+++ public/react/Configure/SettingsSidebar/components/GoBackButton/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./GoBackButton";
