+++ public/react/Configure/SettingsSidebar/components/IssueSource/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./IssueSource"
