+++ public/react/Configure/SettingsSidebar/components/ReportSettings/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./ReportSettings";
