+++ public/react/Configure/SettingsSidebar/components/TimingCalculation/TimingCalculation.test.tsx
@@ -3,11 +3,11 @@ import { render, screen, within } from "@testing-library/react";
 import { describe, it, vi } from "vitest";
 import userEvent from "@testing-library/user-event";
 
-import routeData from "../../../canjs/routing/route-data.js";
+import routeData from "../../../../../canjs/routing/route-data.js";
 
-import { TimingCalculation } from "./TimingCalculation";
+import TimingCalculation from "./TimingCalculation";
 
-vi.mock("../../../canjs/routing/route-data.js", async () => {
+vi.mock("../../../../../canjs/routing/route-data.js", async () => {
   const mockRouteData = {
     simplifiedIssueHierarchy: [
       {
@@ -60,11 +60,11 @@ vi.mock("../../../canjs/routing/route-data.js", async () => {
       Epic: 'childrenFirstThenParent',
     }
   };
-  const { ObservableObject } = await import("../../../can.js");
+  const { ObservableObject } = await import("../../../../../can.js");
 
   return {
     default: new (ObservableObject as ObjectConstructor)(mockRouteData),
-    RouteData: (await vi.importActual("../../../canjs/routing/route-data.js")).RouteData
+    RouteData: (await vi.importActual("../../../../../canjs/routing/route-data.js")).RouteData
   };
 });
