+++ public/react/Configure/SettingsSidebar/components/TimingCalculation/helpers.ts
