+++ public/react/Configure/SettingsSidebar/components/TimingCalculation/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./TimingCalculation";
