+++ public/react/Configure/SettingsSidebar/components/index.ts
@@ -1,0 +1,5 @@
+export { default as GoBackButton } from "./GoBackButton";
+export { default as Branding } from "./Branding";
+export { default as TimingCalculation } from "./TimingCalculation";
+export { default as IssueSource } from "./IssueSource"
+export { default as ReportSettings } from "./ReportSettings";
