--- public/react/Configure/TimingCalculation/index.ts
@@ -1,1 +1,0 @@
-export * from "./TimingCalculation";
